src/backend/requirements.txt
-i https://pypi.org/simple
bcrypt==3.1.7
cffi==1.13.2
click==7.0
cryptography==2.8
ecdsa==0.15
flask-marshmallow==0.10.1
flask-sqlalchemy==2.4.1
flask==1.1.1
itsdangerous==1.1.0
jinja2==2.10.3
markupsafe==1.1.1
marshmallow-sqlalchemy==0.21.0
marshmallow==3.3.0
pyasn1==0.4.8
pycparser==2.19
python-jose[cryptography]==3.1.0
pyyaml==5.3
rsa==4.0
six==1.14.0
sqlalchemy==1.3.13
webargs==5.5.2
werkzeug==0.16.0
